agent/api/serializers.py
from rest_framework import serializers

class MessageInputSerializer(serializers.Serializer):
    message = serializers.CharField(required=True)
    session_id = serializers.CharField(required=False) 